use provisionable mixin for resources, remove unused dashboard template
