--- app/templates/layouts/dashboard2.hbs
@@ -1,24 +1,0 @@
-<header class="layout-header">
-  {{partial "header"}}
-  {{#unless session.currentUser.verified}}
-    {{verification-code-reset}}
-  {{/unless}}
-
-  {{#if flashMessages.queue}}
-    <div class="notifications">
-      {{#each flashMessages.queue as |flash|}}
-        {{flash-message flash=flash}}
-      {{/each}}
-    </div>
-  {{/if}}
-</header>
-
-<div class="layout-body">
-  {{yield}}
-</div>
-
-<footer class="layout-footer">
-  {{partial "nav-footer"}}
-</footer>
-
-
